Add UX improvements to the TasksTable
* Incorporate alerts_count from Tasks new API spec

* Add warning icon to tasks with alerts

* Sort according to start_time instead of end_time

* Remove "delete" and "run task again" buttons

Since "run task again" in original tasks reflected the task type, it
does not make sense here - we only use 1 type of task per
preupgrade/upgrade. It would make sense to run the task with the same
systems, but that is IMO outside the scope of MVP, and would need more
though process going into it.

"Delete" button could be used, but I think we should allow users a
better way to delete things - like mass checkbox deletes. There is
little benefit in having a delete button and going one-by-one, better to
keep the whole history at this point IMO.